feat: extract shared HMR types to pack-shared
