+++ packages/pack-shared/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./config";
+export * from "./hmr";
 export { formatIssue, handleIssues, Issue } from "./issue";
 export { decodeMagicIdentifier } from "./magicIdentifier";
 export { renderStyledStringToErrorAnsi } from "./styledString";
