+++ packages/utoo-web/package.json
@@ -26,7 +26,7 @@
     "prepublishOnly": "turbo run build --filter=@utoo/web"
   },
   "dependencies": {
-    "@utoo/pack-shared": "^1.1.12",
+    "@utoo/pack-shared": "*",
     "comlink": "^4.4.2",
     "systemjs": "^6.15.1"
   },
